Sample diagnostics correctly writing on GPU and being readback to CPU and UI display

+++ src/Shaders/Common.hlsl
@@ -234,7 +234,7 @@ int DiagnosticSlot(uint sampleIndex)
 
 struct DiagnosticDetails
 {
-    float2 gSampledDiagnosticPixel;
+    uint2 gSampledDiagnosticPixel;
     uint isDiagnosticPixel;
     uint gSampleIndex;
 };
